+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -4,13 +4,17 @@ import com.sun.jdi.InternalException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.BadRequestException;
+import ru.yandex.practicum.filmorate.exception.ConflictException;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
 
+import java.time.LocalDate;
 import java.util.Arrays;
-import java.util.Collection;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @Service
@@ -19,12 +23,21 @@ import java.util.stream.Collectors;
 public class UserService {
 
     private final UserStorage userStorage;
+    private final Map<Integer, User> users = new HashMap<>();
+
 
     public User create(User user) {
+        throwIfUserPrintWrongInfo(user);
+        throwIfUserAlreadyExist(user);
         return userStorage.create(user);
     }
 
     public User update(User user) {
+        throwIfUserPrintWrongInfo(user);
+        if (!users.containsKey(user.getId())) {
+            throw new NotFoundException("HTTP ERROR 404: Невозможно обновить данные, так как пользователя не существует");
+        }
+        throwIfUserAlreadyExist(user);
         return userStorage.update(user);
     }
 
@@ -33,75 +46,102 @@ public class UserService {
         return userStorage.findAll();
     }
 
-    public User getById(int id) {
-        if (!userStorage.getAll().containsKey(id)) {
-            throw new NotFoundException("HTTP ERROR 404: Пользователь не найден");
-        }
+    public User findById(int id) {
+        userStorage.isExist(id);
         log.info("Пользователь с id: '{}' отправлен", id);
-        return userStorage.getById(id);
+        return userStorage.findById(id);
     }
 
     public User deleteById(int id) {
-        if (!userStorage.getAll().containsKey(id)) {
-            throw new NotFoundException("HTTP ERROR 404: Пользователь не найден. Невозможно удалить неизветсного пользователя");
-        }
+        userStorage.isExist(id);
         log.info("Пользователь с id: '{}' удален", id);
         return userStorage.deleteById(id);
     }
 
     public List<User> addFriendship(int firstId, int secondId) {
-        if (!userStorage.getAll().containsKey(firstId) || !userStorage.getAll().containsKey(secondId)) {
-            throw new NotFoundException(String.format("Пользователя с id: %d или с id: %d не существует", firstId, secondId));
-        }
-        if (userStorage.getById(firstId).getFriends().contains(secondId)) {
+        userStorage.isExist(firstId);
+        userStorage.isExist(secondId);
+        if (userStorage.findById(firstId).getFriends().contains(secondId)) {
             throw new InternalException("Пользователи уже и так являются друзьями");
         }
-        userStorage.getById(firstId).getFriends().add(secondId);
-        userStorage.getById(secondId).getFriends().add(firstId);
+        userStorage.findById(firstId).getFriends().add(secondId);
+        userStorage.findById(secondId).getFriends().add(firstId);
         log.info("Пользователи: '{}' и '{}' теперь являются друзьями :)",
-                userStorage.getById(firstId).getName(),
-                userStorage.getById(secondId).getName());
-        return Arrays.asList(userStorage.getById(firstId), userStorage.getById(secondId));
+                userStorage.findById(firstId).getName(),
+                userStorage.findById(secondId).getName());
+        return Arrays.asList(userStorage.findById(firstId), userStorage.findById(secondId));
     }
 
     public List<User> removeFriendship(int firstId, int secondId) {
-        if (!userStorage.getAll().containsKey(firstId) || !userStorage.getAll().containsKey(secondId)) {
-            throw new NotFoundException(String.format("Пользователя с id: %d или с id: %d не существует", firstId, secondId));
-        }
-        if (!userStorage.getById(firstId).getFriends().contains(secondId)) {
+        userStorage.isExist(firstId);
+        userStorage.isExist(secondId);
+        if (!userStorage.findById(firstId).getFriends().contains(secondId)) {
             throw new InternalException("Пользователи не являются друзьями");
         }
-        userStorage.getById(firstId).getFriends().remove(secondId);
-        userStorage.getById(secondId).getFriends().remove(firstId);
+        userStorage.findById(firstId).getFriends().remove(secondId);
+        userStorage.findById(secondId).getFriends().remove(firstId);
         log.info("Пользователи: '{}' и '{}' больше не друзья :(",
-                userStorage.getById(firstId).getName(),
-                userStorage.getById(secondId).getName());
-        return Arrays.asList(userStorage.getById(firstId), userStorage.getById(secondId));
+                userStorage.findById(firstId).getName(),
+                userStorage.findById(secondId).getName());
+        return Arrays.asList(userStorage.findById(firstId), userStorage.findById(secondId));
     }
 
     public List<User> getFriendsListById(int id) {
-        if (!userStorage.getAll().containsKey(id)) {
-            throw new NotFoundException("HTTP ERROR 404: Невозможно получить список друзей пользователя, " +
-                    "так как пользователь не найден :(");
-        }
+        userStorage.isExist(id);
         log.info("Успех! Запрос получения списка друзей пользователя '{}' выполнен успешно :)",
-                userStorage.getById(id).getName());
-        return userStorage.getById(id).getFriends().stream()
-                .map(userStorage::getById)
+                userStorage.findById(id).getName());
+        return userStorage.findById(id).getFriends().stream()
+                .map(userStorage::findById)
                 .collect(Collectors.toList());
     }
 
     public List<User> getCommonFriendsList(int firstId, int secondId) {
-        if (!userStorage.getAll().containsKey(firstId) || !userStorage.getAll().containsKey(secondId)) {
-            throw new NotFoundException(String.format("Пользователь с id: %d или с id: %d не существует :(", firstId, secondId));
-        }
-        User firstUser = userStorage.getById(firstId);
-        User secondUser = userStorage.getById(secondId);
+        userStorage.isExist(firstId);
+        userStorage.isExist(secondId);
+        User firstUser = userStorage.findById(firstId);
+        User secondUser = userStorage.findById(secondId);
         log.info("Список общих друзей пользователей: '{}' и '{}' успешено отправлен",
                 firstUser.getName(), secondUser.getName());
         return firstUser.getFriends().stream()
                 .filter(friendId -> secondUser.getFriends().contains(friendId))
-                .map(userStorage::getById)
+                .map(userStorage::findById)
                 .collect(Collectors.toList());
+    }
+
+    void throwIfUserPrintWrongInfo(User user) {
+
+        if (user.getLogin().contains(" ") || user.getLogin().isBlank()) {
+            log.warn("Введенный Логин пользователя: '{}'", user.getLogin());
+            throw new BadRequestException("HTTP ERROR 400: Логин не может быть пустым");
+        }
+
+        if (user.getName() == null || user.getName().equals("")) {
+            user.setName(user.getLogin());
+            log.warn("Не заполнено Имя пользователя заменено на Логин: '{}'", user.getName());
+        }
+
+        if (user.getBirthday().isAfter(LocalDate.now())) {
+            log.warn("Указанная Дата рождения: '{}'", user.getBirthday());
+            throw new BadRequestException("HTTP ERROR 400: Дата рождения не может быть в будущем");
+        }
+
+        if (user.getEmail().isBlank() || user.getEmail() == null || user.getEmail().equals(" ")) {
+            log.warn("Введенный Email пользователя: '{}'", user.getEmail());
+            throw new BadRequestException("HTTP ERROR 400: Email не может быть пустым");
+        }
+    }
+
+    private void throwIfUserAlreadyExist(User userToAdd) {
+        boolean exists = users.values().stream()
+                .anyMatch(user -> isAlreadyExist(userToAdd, user));
+        if (exists) {
+            log.warn("Введенный Email пользователя: '{}'", userToAdd);
+            throw new ConflictException("HTTP ERROR 409: Пользователь с таким Email или логином уже существует");
+        }
+    }
+
+    private boolean isAlreadyExist(User userToAdd, User user) {
+        return userToAdd.getLogin().equals(user.getLogin()) ||
+                userToAdd.getEmail().equals(user.getEmail());
     }
 }
