+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -3,7 +3,6 @@ package ru.yandex.practicum.filmorate.storage;
 import ru.yandex.practicum.filmorate.model.Film;
 
 import java.util.List;
-import java.util.Map;
 
 public interface FilmStorage {
 
@@ -11,9 +10,10 @@ public interface FilmStorage {
 
     Film update(Film film);
 
-    Film getById(int id);
+    Film findById(int id);
     Film deleteById(int id);
 
     List<Film> findAll();
-    Map<Integer, Film> getAll();
+
+    Film isExist (int id);
 }
