+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryUserStorage.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exception.BadRequestException;
 import ru.yandex.practicum.filmorate.exception.ConflictException;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.time.LocalDate;
@@ -19,8 +20,6 @@ public class InMemoryUserStorage implements UserStorage {
 
     @Override
     public User create(User user) {
-        throwIfUserPrintWrongInfo(user);
-        throwIfUserAlreadyExist(user);
         user.setId(id++);
         users.put(user.getId(), user);
         log.info("Пользователь успешно добавлен с логином: {}, email: {}", user.getLogin(), user.getEmail());
@@ -29,11 +28,6 @@ public class InMemoryUserStorage implements UserStorage {
 
     @Override
     public User update(User user) {
-        throwIfUserPrintWrongInfo(user);
-        if (!users.containsKey(user.getId())) {
-            throw new NotFoundException("HTTP ERROR 404: Невозможно обновить данные, так как пользователя не существует");
-        }
-        throwIfUserAlreadyExist(user);
         users.put(user.getId(), user);
         log.info("Данные пользователя с id: {}, логином: {} успешно обновлена", user.getId(), user.getLogin());
         return user;
@@ -45,12 +39,7 @@ public class InMemoryUserStorage implements UserStorage {
     }
 
     @Override
-    public Map<Integer, User> getAll() {
-        return users;
-    }
-
-    @Override
-    public User getById(int id) {
+    public User findById(int id) {
         return users.get(id);
     }
 
@@ -61,42 +50,12 @@ public class InMemoryUserStorage implements UserStorage {
         return user;
     }
 
-    void throwIfUserPrintWrongInfo(User user) {
-
-        if (user.getLogin().contains(" ") || user.getLogin().isBlank()) {
-            log.warn("Введенный Логин пользователя: '{}'", user.getLogin());
-            throw new BadRequestException("HTTP ERROR 400: Логин не может быть пустым");
+    @Override
+    public User isExist(int id) {
+        User user = users.get(id);
+        if (!users.containsKey(id)) {
+            throw new NotFoundException("HTTP ERROR 404: Пользователь не найден");
         }
-
-        if (user.getName() == null || user.getName().equals("")) {
-            user.setName(user.getLogin());
-            log.warn("Не заполнено Имя пользователя заменено на Логин: '{}'", user.getName());
-        }
-
-        if (user.getBirthday().isAfter(LocalDate.now())) {
-            log.warn("Указанная Дата рождения: '{}'", user.getBirthday());
-            throw new BadRequestException("HTTP ERROR 400: Дата рождения не может быть в будущем");
-        }
-
-        if (user.getEmail().isBlank() || user.getEmail() == null || user.getEmail().equals(" ")) {
-            log.warn("Введенный Email пользователя: '{}'", user.getEmail());
-            throw new BadRequestException("HTTP ERROR 400: Email не может быть пустым");
-        }
+        return user;
     }
-
-    private void throwIfUserAlreadyExist(User userToAdd) {
-        boolean exists = users.values().stream()
-                .anyMatch(user -> isAlreadyExist(userToAdd, user));
-        if (exists) {
-            log.warn("Введенный Email пользователя: '{}'", userToAdd);
-            throw new ConflictException("HTTP ERROR 409: Пользователь с таким Email или логином уже существует");
-        }
-    }
-
-    private boolean isAlreadyExist(User userToAdd, User user) {
-        return userToAdd.getLogin().equals(user.getLogin()) ||
-                userToAdd.getEmail().equals(user.getEmail());
-
-    }
-
 }
